Carry mouse positions as floats, and hit-test scroll events where the wheel moved
- `poll_events` no longer truncates SDL's converted coordinates: under a scaled viewport a window pixel is a fraction of a logical unit, so a click at 10.75 was delivered as 10 and missed a `Rectangle` at `x: 10.5`. `R2D_QueuedEvent` and `R2D_Mouse` store floats, `drain_events` and the `@mouse_x`/`@mouse_y` sync push Floats, and `MouseEvent#x`/`#y` and `mouse_x`/`mouse_y` are Floats on every path.
- Wheel events queue `e.wheel.mouse_x`/`mouse_y`, converted with the rest of the event, and `handle_mouse_scroll` dispatches to the object under that position instead of the polled position, which by then can be over another object. Window-level `:mouse_scroll` events carry the position too.
- Examples that divide mouse coordinates into grid cells (`astar`, `falling_sand`, `game_of_life`) `floor` the quotient now that it is a Float.
- `USAGE.md`: positions are floats; a `:mouse_scroll` event's position is where the wheel moved.

+++ ext/ruby2d/ruby2d.h
@@ -465,8 +465,8 @@ typedef struct {
   int button;
   int direction;
   int axis;
-  int x;
-  int y;
+  float x;
+  float y;
   float delta_x;
   float delta_y;
   int value;
@@ -481,10 +481,12 @@ typedef struct {
   float a;
 } R2D_Color;
 
-// R2D_Mouse
+// R2D_Mouse — cursor position in logical render coordinates. Kept as floats:
+// under a scaled viewport a window pixel maps to a fraction of a logical
+// unit, and truncating would push hit-tests off fractional object edges.
 typedef struct {
-  int x;
-  int y;
+  float x;
+  float y;
 } R2D_Mouse;
 
 // R2D_Viewport
